Added middleware

+++ controllers/goalController.js
@@ -5,13 +5,16 @@ const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
 
 
+//Bring userModel
+const User = require('../models/userModels')
+
 // @desc   Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals =  async (req, res) => {
 
     //fetch from DB
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
         
     res.status(200).json(goals)
 }
@@ -30,7 +33,8 @@ const addGoals =  asyncHandler(async (req, res) => {
 
         //Save to DB
         const goal = await Goal.create({
-            text: req.body.text
+            text: req.body.text,
+            user: req.user.id
         })
     
     res.status(200).json({
@@ -51,6 +55,19 @@ const updateGoals =  asyncHandler(async  (req, res) => {
             message: `Goal with ${req.params.id} not found`
         })
     } 
+
+    const user =await User.findById(req.user.id)
+    //check user
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged users matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(400)
+        throw new Error('User not authorised')
+    }
     
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -74,9 +91,24 @@ const deleteGoals =  asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400) 
         throw new Error('Goal not found')
-    }  
-        await goal.deleteOne();
-        res.status(200).json({message: `Goals deleted ${req.params.id}`})
+    } 
+
+    const user =await User.findById(req.user.id)
+    //check user
+    if(!user){
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged users matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(400)
+        throw new Error('User not authorised')
+    }
+    
+    
+    await goal.deleteOne();
+    res.status(200).json({message: `Goals deleted ${req.params.id}`})
    
 
 })
